pc/requirements.txt
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
functools32==3.2.3.post2
idna==2.7
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
matplotlib==1.5.1
nbformat==4.4.0
numpy==1.14.0
paho-mqtt==1.3.1
pandas==0.22.0
plotly==3.6.1
pyparsing==2.2.0
pyserial==3.2.1
python-dateutil==2.6.1
pytz==2017.3
requests==2.19.1
retrying==1.3.3
scipy==1.2.1
six==1.11.0
traitlets==4.3.2
urllib3==1.23
wincertstore==0.2
wxPython==3.0.2.0
